backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
MongoEngine==0.29.0
Flask_Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
Flask-Talisman==1.0.0